:recycle: refactor: exception invocation
