+++ src/main/java/com/honlife/core/app/model/loginLog/service/LoginLogService.java
@@ -1,5 +1,6 @@
 package com.honlife.core.app.model.loginLog.service;
 
+import com.honlife.core.infra.response.ResponseCode;
 import jakarta.transaction.Transactional;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
@@ -11,7 +12,7 @@ import com.honlife.core.app.model.loginLog.dto.LoginLogDTO;
 import com.honlife.core.app.model.loginLog.repos.LoginLogRepository;
 import com.honlife.core.app.model.member.domain.Member;
 import com.honlife.core.app.model.member.repos.MemberRepository;
-import com.honlife.core.infra.util.NotFoundException;
+import com.honlife.core.infra.error.exceptions.NotFoundException;
 
 
 @Service
@@ -31,7 +32,7 @@ public class LoginLogService {
     public LoginLogDTO get(final Long id) {
         return loginLogRepository.findById(id)
             .map(loginLog -> mapToDTO(loginLog, new LoginLogDTO()))
-            .orElseThrow(NotFoundException::new);
+            .orElseThrow(() -> new NotFoundException(ResponseCode.NOT_FOUND));
     }
 
     public Long create(final LoginLogDTO loginLogDTO) {
@@ -42,7 +43,7 @@ public class LoginLogService {
 
     public void update(final Long id, final LoginLogDTO loginLogDTO) {
         final LoginLog loginLog = loginLogRepository.findById(id)
-            .orElseThrow(NotFoundException::new);
+            .orElseThrow(() -> new NotFoundException(ResponseCode.NOT_FOUND));
         mapToEntity(loginLogDTO, loginLog);
         loginLogRepository.save(loginLog);
     }
@@ -61,7 +62,7 @@ public class LoginLogService {
     private LoginLog mapToEntity(final LoginLogDTO loginLogDTO, final LoginLog loginLog) {
         loginLog.setTime(loginLogDTO.getTime());
         final Member member = loginLogDTO.getMember() == null ? null : memberRepository.findById(loginLogDTO.getMember())
-            .orElseThrow(() -> new NotFoundException("member not found"));
+            .orElseThrow(() -> new NotFoundException(ResponseCode.NOT_FOUND_MEMBER));
         loginLog.setMember(member);
         return loginLog;
     }
