+++ src/main/java/com/honlife/core/app/model/member/service/MemberItemService.java
@@ -1,6 +1,7 @@
 package com.honlife.core.app.model.member.service;
 
 import com.honlife.core.app.model.category.domain.Category;
+import com.honlife.core.infra.response.ResponseCode;
 import java.util.List;
 import org.springframework.data.domain.Sort;
 import org.springframework.security.core.userdetails.UserDetails;
@@ -12,7 +13,7 @@ import com.honlife.core.app.model.member.domain.MemberItem;
 import com.honlife.core.app.model.member.model.MemberItemDTO;
 import com.honlife.core.app.model.member.repos.MemberItemRepository;
 import com.honlife.core.app.model.member.repos.MemberRepository;
-import com.honlife.core.infra.util.NotFoundException;
+import com.honlife.core.infra.error.exceptions.NotFoundException;
 import org.springframework.transaction.annotation.Transactional;
 
 
@@ -40,7 +41,7 @@ public class MemberItemService {
     public MemberItemDTO get(final Long id) {
         return memberItemRepository.findById(id)
                 .map(memberItem -> mapToDTO(memberItem, new MemberItemDTO()))
-                .orElseThrow(NotFoundException::new);
+                .orElseThrow(() -> new NotFoundException(ResponseCode.NOT_FOUND_ITEM));
     }
 
     public Long create(final MemberItemDTO memberItemDTO) {
@@ -51,7 +52,7 @@ public class MemberItemService {
 
     public void update(final Long id, final MemberItemDTO memberItemDTO) {
         final MemberItem memberItem = memberItemRepository.findById(id)
-                .orElseThrow(NotFoundException::new);
+                .orElseThrow(() -> new NotFoundException(ResponseCode.NOT_FOUND_ITEM));
         mapToEntity(memberItemDTO, memberItem);
         memberItemRepository.save(memberItem);
     }
@@ -77,10 +78,10 @@ public class MemberItemService {
         memberItem.setIsActive(memberItemDTO.getIsActive());
         memberItem.setIsEquipped(memberItemDTO.getIsEquipped());
         final Member member = memberItemDTO.getMember() == null ? null : memberRepository.findById(memberItemDTO.getMember())
-                .orElseThrow(() -> new NotFoundException("member not found"));
+                .orElseThrow(() -> new NotFoundException(ResponseCode.NOT_FOUND_MEMBER));
         memberItem.setMember(member);
         final Item item = memberItemDTO.getItem() == null ? null : itemRepository.findById(memberItemDTO.getItem())
-                .orElseThrow(() -> new NotFoundException("item not found"));
+                .orElseThrow(() -> new NotFoundException(ResponseCode.NOT_FOUND_ITEM));
         memberItem.setItem(item);
         return memberItem;
     }
